@@ -8,55 +9,53 @@
     "test": "jest --coverage=false",
     "test:ci": "jest --ci --coverage=true",
     "lint": "npm run lint:ts && npm run lint:styles",
-    "lint:ts": "npx ng lint --fix",
+    "lint:ts": "eslint --fix",
     "lint:styles": "stylelint --fix src/**/*.scss",
-    "audit:ci": "npm audit --json|npm-audit-reporter-teamcity",
     "nswag": "node_modules\\.bin\\nswag run nswag.json",
-    "codegen:run": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net8.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features",
-    "codegen:run-i18n": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net8.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features --enable-i18n true",
+    "codegen:run": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features",
+    "codegen:run-i18n": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features --enable-i18n true",
     "codegen:install": "dotnet tool install enigmatry.entry.codegeneration.tools -g",
     "codegen:update": "dotnet tool update enigmatry.entry.codegeneration.tools -g",
     "i18n": "ng extract-i18n --format json --output-path src/i18n --out-file messages.en.json & sortjson src/i18n"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.13",
-    "@angular/cdk": "^18.2.14",
-    "@angular/common": "^18.2.13",
-    "@angular/compiler": "^18.2.13",
-    "@angular/core": "^18.2.13",
-    "@angular/forms": "^18.2.13",
-    "@angular/material": "^18.2.14",
-    "@angular/material-date-fns-adapter": "^18.2.14",
-    "@angular/platform-browser": "^18.2.13",
-    "@angular/platform-browser-dynamic": "^18.2.13",
-    "@angular/router": "^18.2.13",
+    "@angular/animations": "^19.2.10",
+    "@angular/cdk": "^19.2.15",
+    "@angular/common": "^19.2.10",
+    "@angular/compiler": "^19.2.10",
+    "@angular/core": "^19.2.10",
+    "@angular/forms": "^19.2.10",
+    "@angular/material": "^19.2.15",
+    "@angular/material-date-fns-adapter": "^19.2.15",
+    "@angular/platform-browser": "^19.2.10",
+    "@angular/platform-browser-dynamic": "^19.2.10",
+    "@angular/router": "^19.2.10",
     "@azure/msal-browser": "3.2.0",
-    "@enigmatry/entry-components": "18.0.0",
-    "@enigmatry/entry-form": "18.0.0",
+    "@enigmatry/entry-components": "19.0.0",
+    "@enigmatry/entry-form": "19.0.0",
     "@microsoft/applicationinsights-web": "3.3.4",
-    "@ngx-formly/core": "6.3.11",
-    "@ngx-formly/material": "6.3.11",
+    "@ngx-formly/core": "6.3.12",
+    "@ngx-formly/material": "6.3.12",
     "date-fns": "^3.6.0",
-    "rxjs": "^7.8.1",
+    "rxjs": "^7.8.2",
     "tslib": "^2.8.1",
-    "zone.js": "~0.14.10"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular/build": "^18.2.12",
-    "@angular/cli": "^18.2.12",
-    "@angular/compiler-cli": "^18.2.13",
-    "@angular/localize": "^18.2.13",
-    "@enigmatry/eslint-config": "18.0.0",
-    "@enigmatry/scss-foundation": "18.0.0",
-    "@enigmatry/stylelint-config": "18.0.0",
+    "@angular/build": "^19.2.11",
+    "@angular/cli": "^19.2.11",
+    "@angular/compiler-cli": "^19.2.10",
+    "@angular/localize": "^19.2.10",
+    "@enigmatry/eslint-config": "19.0.0",
+    "@enigmatry/scss-foundation": "19.0.0",
+    "@enigmatry/stylelint-config": "19.0.0",
     "@types/jest": "29.5.14",
     "jest": "29.7.0",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "14.2.0",
+    "jest-preset-angular": "14.5.5",
     "json-sort-cli": "^4.0.3",
-    "npm-audit-reporter-teamcity": "0.4.0",
-    "nswag": "^14.2.0",
-    "typescript": "5.5.4"
+    "nswag": "^14.4.0",
+    "typescript": "5.8.3"
   }
 }
