@@ -11,6 +11,14 @@ namespace Enigmatry.Entry.Blueprint.CodeGeneration.Setup.Features.Products;
 [UsedImplicitly]
 public class ProductEditComponentConfiguration : IFormComponentConfiguration<GetProductDetails.Response>
 {
+    internal const string AmountLabel = "products.amount";
+    internal const string PriceLabel = "products.price";
+    private const string TypeLabel = "products.product-edit.type.label";
+    private const string DescriptionLabel = "products.product-edit.description.label";
+    private const string DiscountLabel = "products.product-edit.discount.label";
+    private const string HasDiscountLabel = "products.product-edit.has-discount.label";
+    private const string FreeShippingLabel = "products.product-edit.free-shipping.label";
+
     public void Configure(FormComponentBuilder<GetProductDetails.Response> builder)
     {
         // Configuring component:
@@ -29,23 +37,23 @@ public class ProductEditComponentConfiguration : IFormComponentConfiguration<Get
             .WithValidators("productCodeIsUnique");
 
         builder.AutocompleteFormControl(x => x.Type)
-            .WithOptions(options => options.WithFixedValues<ProductType>().WithSortKey("displayName"));
+            .WithOptions(options => options.WithFixedValues<ProductType>().WithSortKey("displayName"))
+            .WithTranslationId(TypeLabel);
 
         builder.TextareaFormControl(x => x.Description)
-            .WithRows(2);
+            .WithRows(2)
+            .WithTranslationId(DescriptionLabel);
 
         builder.FormControl(x => x.Price)
             .WithLabel("Price per unit")
-            .WithLabelTranslationId(ProductTranslationId.Price)
             .WithPlaceholder("Price per unit")
-            .WithPlaceholderTranslationId(ProductTranslationId.Price)
+            .WithTranslationId(PriceLabel)
             .WithFormat(new CurrencyPropertyFormatter().WithCurrencyCode("EUR").WithDisplay("€"));
 
         builder.FormControl(x => x.Amount)
             .WithLabel("Units")
-            .WithLabelTranslationId(ProductTranslationId.Amount)
             .WithPlaceholder("Units")
-            .WithPlaceholderTranslationId(ProductTranslationId.Amount);
+            .WithTranslationId(AmountLabel);
 
         builder.FormControl(x => x.ContactEmail)
             .WithPlaceholder("Contact person email address");
@@ -62,12 +70,15 @@ public class ProductEditComponentConfiguration : IFormComponentConfiguration<Get
         builder.FormControl(x => x.ExpiresOn)
             .WithPlaceholder("Product expiration date, if any");
 
-        builder.FormControl(x => x.FreeShipping);
+        builder.FormControl(x => x.FreeShipping)
+            .WithTranslationId(FreeShippingLabel);
 
-        builder.FormControl(x => x.HasDiscount);
+        builder.FormControl(x => x.HasDiscount)
+            .WithTranslationId(HasDiscountLabel);
 
         builder.FormControl(x => x.Discount)
-            .WithFormat(new PercentPropertyFormatter().WithDigitsInfo("1.2-2"));
+            .WithFormat(new PercentPropertyFormatter().WithDigitsInfo("1.2-2"))
+            .WithTranslationId(DiscountLabel);
 
         builder.ButtonFormControl("ResetFormBtn")
             .WithLabel(String.Empty)
