@@ -1,6 +1,6 @@
 //----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v14.2.0.0 (NJsonSchema v11.1.0.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v14.4.0.0 (NJsonSchema v11.3.2.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 
@@ -844,7 +844,7 @@ export class PagedResponseOfGetUsersResponseItem implements IPagedResponseOfGetU
         if (Array.isArray(this.items)) {
             data["items"] = [];
             for (let item of this.items)
-                data["items"].push(item.toJSON());
+                data["items"].push(item ? item.toJSON() : <any>undefined);
         }
         data["totalCount"] = this.totalCount;
         data["pageNumber"] = this.pageNumber;
@@ -1286,7 +1286,7 @@ export class PagedResponseOfGetProductsResponseItem implements IPagedResponseOfG
         if (Array.isArray(this.items)) {
             data["items"] = [];
             for (let item of this.items)
-                data["items"].push(item.toJSON());
+                data["items"].push(item ? item.toJSON() : <any>undefined);
         }
         data["totalCount"] = this.totalCount;
         data["pageNumber"] = this.pageNumber;
