@@ -5,7 +5,5 @@ export class UserProfile extends GetUserProfileResponse {
     return this.permissions?.includes(permission) ?? false;
   }
 
-  static fromResponse(response: IGetUserProfileResponse | null) {
-    return response ? new UserProfile(response) : null;
-  }
+  static fromResponse = (response: IGetUserProfileResponse | null) => response ? new UserProfile(response) : null;
 }
