@@ -9,29 +9,29 @@ export abstract class BaseListComponent {
   constructor(protected router: Router, protected activatedRoute: ActivatedRoute) { }
 
   // Handle search filters change
-  searchFilterChange(searchParams: SearchFilterParams) {
+  readonly searchFilterChange = async(searchParams: SearchFilterParams) => {
     this.query.searchFilterChange(searchParams);
-    this.updateRouteQueryString(this.query.getRouteQueryParams());
-  }
+    await this.updateRouteQueryString(this.query.getRouteQueryParams());
+  };
 
   // Handle page change
-  pageChange(page: PageEvent): void {
+  readonly pageChange = async(page: PageEvent) => {
     this.query.pageChange(page);
-    this.updateRouteQueryString(this.query.getRouteQueryParams());
-  }
+    await this.updateRouteQueryString(this.query.getRouteQueryParams());
+  };
 
   // Handle sort change
-  sortChange(sort: SortEvent): void {
+  readonly sortChange = async(sort: SortEvent) => {
     this.query.sortChange(sort);
-    this.updateRouteQueryString(this.query.getRouteQueryParams());
-  }
+    await this.updateRouteQueryString(this.query.getRouteQueryParams());
+  };
 
   // Appends new query params to the current route
-  private updateRouteQueryString(queryParams: Params): void {
-    this.router.navigate([], {
+  private readonly updateRouteQueryString = async(queryParams: Params) => {
+    await this.router.navigate([], {
       relativeTo: this.activatedRoute,
       queryParams,
       queryParamsHandling: 'merge'
     });
-  }
+  };
 }
